Viewer: quantize PM2.5 to R8 in worker; upload as r8unorm to GPU

+++ viewer/src/useForecast.ts
@@ -15,7 +15,10 @@ export type ForecastMeta = {
 };
 
 export type Frame = {
-  data: Float32Array;
+  /** PM2.5 quantized to 0-255 against PM_MAX in the worker. Multiply by
+   *  PM_MAX/255 to recover µg/m³ for CPU-side consumers; the GPU samples
+   *  the r8unorm texture directly into the colormap LUT. */
+  data: Uint8Array;
   maxPm25: number;
   initTime: number;
 };
